allow controlling the behavior wrt symlinks
Add an option ``-8`` with three possible arguments:

    - none: process all symlinks (default);
    - all: ignore all symlinks;
    - absolute: ignore symlinks to absolute paths

This works both ways, when creating a cpio archive and when
extracting the content of an archives. Links skipped thusly are
indicated with a message on stderr.

The command line flag needs to be ``-8`` since the present
options already exhaust the entire English alphabet plus all
digits except for eight.

Signed-off-by: Philipp Gesang <philipp.gesang@intra2net.com>