Support PREFIX and DESTDIR Makefile variables
Let the user specify PREFIX to indicate where the program and manpage
should be installed. This changes the default installation location of
the manpage from /usr/share/man/man1 to /usr/local/share/man/man1 but
resolves the previous confusion where the program was installed to
prefix /usr/local but the manpage was installed to prefix /usr.

By setting DESTDIR a user can stage the installation to a temporary
location rather than installing directly to PREFIX.